cython: Ported resample extension to Cython, and add run-time compilation attempt for platforms that have not had the extension precompiled.
The port is a straight copy-paste of resample4 from __main__.py with replacements
of len() and [-1] with the known length N of the arrays and N - 1 respectively.

Compilation no longer has a numpy dendendency since Cython has its own arrays.

The Cython extension does not hold the GIL, whereas I believe the C extension did.