Require pre-release pyqtgraph on Python 3.8
I thought depending on labscript_utils which depends on the newer
pyqtgraph would be sufficient, but it's not - if runviewer is processed
by pip first, pip does not install the pre-release version even though
labscript_utils requires is. Maybe a bug in pip, or known deficiency in
its dependency resolution, but this fixes it.