Improved performance of resample thread
While the resapmle algorithm released the GIL, we did not take advantage of the fact that this meant we could resample multiple channels in parallel.
A ThreadPool is now created to resample channels in parallel.
The code was also optimised to remove logging lines and reduce the number of calls in `inmain`, thus limiting the contention with pyqtgraph redraws and mouse move events.